ci: exclude galoy repo for test-integration.yml

+++ ci/tasks/bump-shared-files.sh
@@ -31,7 +31,10 @@ if [[ ! -f ./typos.toml ]]; then
 fi
 
 if [[ -f ./docker-compose.yml ]]; then
-  cp ./ci/vendor/actions/test-integration.yml ./.github/workflows
+  # exclude galoy repo
+  if [[ ! -f .github/workflows/integration-test.yml ]]; then
+    cp ./ci/vendor/actions/test-integration.yml ./.github/workflows
+  fi
 fi
 
 if [[ -z $(git config --global user.email) ]]; then
